# docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.Model:
    properties:
      connection_id:
        type: string
      id:
        type: string
      price_input:
        type: number
      price_output:
        type: number
      provider_model_id:
        type: string
      proxy_model_id:
        type: string
      thinking:
        type: boolean
      tools_usage:
        type: boolean
    type: object
...
  api.Provider:
    properties:
      base_url:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Proxy server for Generative AI models.
  license:
    name: Apache 2.0
  title: GenAI Proxy API
  version: "1.0"
paths:
...
  /api/models:
    get:
      consumes:
      - application/json
      description: List all available models.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all models
      tags:
      - Models
    post:
      consumes:
      - application/json
      description: Create a new model for a specific connection.
      parameters:
      - description: Model details
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api.Model'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new model
      tags:
      - Models
  /api/models/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an existing model by ID.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete a model
      tags:
      - Models
  /api/providers:
    get:
      consumes:
      - application/json
      description: List all available providers.
      parameters:
      - description: Provider details
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/api.Provider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Provider'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create a new provider.
      parameters:
      - description: Provider details
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/api.Provider'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Provider'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new provider
      tags:
      - Providers
  /api/providers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a provider by ID.
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a provider
      tags:
      - Providers
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter your username and password to get a token.
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"